revert Dataset component move - not just used in catalogue
